requirements.txt
python-telegram-bot
yfinance~=1.2.0
pandas
apscheduler~=3.11.2
python-dotenv~=1.2.2
dotenv~=0.9.9
pytz~=2026.1.post1